requirements.txt
# Chess Engine Dependencies - Phase 1
# ====================================

# Core Chess Library
python-chess>=1.999
    # Provides move generation, board representation, and UCI utilities
    # Documentation: https://python-chess.readthedocs.io/

# Numerical Computing
numpy>=1.24.0
    # Used for tensor operations and board representation
    # Required for Phase 3 neural network integration

# Jupyter and Visualization
jupyter>=1.0.0
    # Interactive notebooks for board visualization
ipython>=8.0.0
    # Enhanced Python REPL for notebooks
matplotlib>=3.7.0
    # Plotting evaluation curves and analysis

# Testing
pytest>=7.0.0
    # Unit testing framework
pytest-cov>=4.0.0
    # Code coverage reporting

# Development Tools
black>=23.0.0

flake8>=6.0.0
    # Linting
mypy>=1.0.0
    # Type checking

h5py>=3.9.0
    # HDF5 file format for efficient dataset storage

stockfish>=3.28.0
    # Python wrapper for Stockfish chess engine

tqdm>=4.66.0
    # Progress bars

pandas>=2.0.0
    # Data analysis

# Neural Network Training
torch>=2.0.0
    # PyTorch for neural network training
torchvision>=0.15.0
    # Vision utilities
tensorboard>=2.14.0
    # Training visualization and logging
